+++ backend/config/dev.exs
@@ -58,7 +58,7 @@ config :backend, Backend.Repo,
 
 config :backend, :crawler,
   status_age_limit_days: 28,
-  status_count_limit: 100,
+  status_count_limit: 1000,
   personal_instance_threshold: 5,
   crawl_interval_mins: 60,
   crawl_workers: 10,